コードレビューを基に修正
