+++ Styles/appStyles/Home/Home.module.scss
@@ -4,7 +4,7 @@
     margin-top: 35px;
 }
 
-.calendar {
+.Calendar {
     width: 90%;
     margin: auto;
     margin-top: 15px;
