+++ Styles/componentStyles/Footer.module.scss
@@ -6,7 +6,7 @@
     bottom: 0;
     display: flex;
     justify-content: space-around;
-    .navBtn {
+    .NavBtn {
         margin-top: 12px;
         text-align: center;
         color: #fff;
