+++ src/app/home/utils/ScheduleViewArray.ts
@@ -1,0 +1,28 @@
+const ScheduleViewArray = [
+    {
+        month: 1,
+        day: 10,
+        TimeStart: '10:00',
+        TimeEnd: '22:30',
+        Title: '直近の予定直近の予定',
+        SubTitle: '直近の予定直近の予定',
+    },
+    {
+        month: 1,
+        day: 11,
+        TimeStart: '11:00',
+        TimeEnd: '23:30',
+        Title: '直近の予定直近の予定2',
+        SubTitle: '直近の予定直近の予定2',
+    },
+    {
+        month: 1,
+        day: 12,
+        TimeStart: '12:00',
+        TimeEnd: '24:30',
+        Title: '直近の予定直近の予定3',
+        SubTitle: '直近の予定直近の予定3',
+    },
+];
+
+export default ScheduleViewArray;
